# Validate
- 기존 코드에서 실제 `Muser`에 대해서도 가짜 `Evaluation`을 만들어내는 오류가 있었습니다.
이를 바로 잡아, 실제 `Muser`에 대해서는 그 `Muser`의 `Evaluation`을 DB에서 꺼내오는 방식으로 바꾸었습니다.

# Optimize
## Timer
- 각 서브루틴의 수행시간을 쉽게 알 수 있도록 타이머 클래스를 정의하고 코드 내에서 사용하였습니다.
  - `timer.start`: 서브루틴의 시작을 알립니다. 메시지를 인자로 전달합니다.
  - `timer.end`: 서브루틴의 끝을 알립니다. 화면에 서브루틴의 수행시간이 표시됩니다.
  - `timer.info`: 화면 상에 정보를 표시합니다. 메시지를 인자로 전달합니다.
  - start-end 쌍은 중첩하여 사용할 수 있습니다. 이 경우에는 화면 상에서 인덴팅이 다르게 표시됩니다.

- 다음과 같이 화면에 표시됩니다.
```
Gather Genres: Finished 0.709 s
Fetch Artist Ids: Finished 0.137 s
Cluster:
  Construct Random Instances:
    Fetch Musics: Finished 62.496 s
    Fetch Real Musers: Finished 0.012 s
    Construct Random Dummy Musers: Finished 3.800 s
    Construct Random Evaluations For Dummy Musers: Finished 22.297 s
  Construct Random Instances: Finished 88.606 s
  Init Clusters: Finished 0.151 s
  Step 1: Finished 0.001 s
  Step 2, 3, 4:
    depth: 2 (200 clusters) [1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1]
    depth: 1 (100 clusters) [2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2]
    depth: 0 (50 clusters) [4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4]
  Step 2, 3, 4: Finished 0.312 s
  Step 5: Finished 2.336 s
  Clusters (Number, Size): [(0, 419), (1, 301), (2, 263), (3, 220), (4, 166), (5, 130), (6, 89), (7, 80), (8, 73), (9, 47), (10, 35), (11, 23), (12, 15), (13, 25), (14, 14), (15, 20), (16, 6), (17, 5), (18, 6), (19, 8), (20, 4), (21, 3), (22, 6), (23, 1), (24, 4), (25, 1), (26, 1), (27, 4), (28, 0), (29, 1), (30, 3), (31, 2), (32, 3), (33, 3), (34, 0), (35, 0), (36, 0), (37, 1), (38, 2), (39, 2), (40, 1), (41, 0), (42, 3), (43, 1), (44, 2), (45, 0), (46, 3), (47, 3), (48, 1), (49, 1)]
  Save Clusters: Finished 3.794 s
Cluster: Finished 95.221 s
```

## 최적화 결과
- `num_users = 1000`
- `num_music = Music.objects.count()`
- `num_evals = 50`
의 경우에 약 4배의 속도 향상을 보였습니다.
또한 음악 정보를 DB에서 꺼내는 데에 약 75%의 시간이 사용되고, 나머지 25%만 `num_users x num_evals`의 크기에 거의 비례하여 시간이 늘어납니다.